# web/frontend/src/client/backend.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Platformatic",
    "description": "This is a service built on top of Platformatic",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/runtimes": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "includeAdmin",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "pid": {
                        "type": "integer"
                      },
                      "cwd": {
                        "type": "string"
                      },
                      "argv": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "uptimeSeconds": {
                        "type": "number"
                      },
                      "execPath": {
                        "type": "string"
                      },
                      "nodeVersion": {
                        "type": "string"
                      },
                      "projectDir": {
                        "type": "string"
                      },
                      "packageName": {
                        "type": "string"
                      },
                      "packageVersion": {
                        "type": "string"
                      },
                      "url": {
                        "type": "string"
                      },
                      "platformaticVersion": {
                        "type": "string"
                      },
                      "selected": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "pid",
                      "cwd",
                      "argv",
                      "uptimeSeconds",
                      "execPath",
                      "nodeVersion",
                      "projectDir",
                      "packageName",
                      "packageVersion",
                      "url",
                      "platformaticVersion",
                      "selected"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/runtimes/{pid}/services": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "pid",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "entrypoint",
                    "production",
                    "services"
                  ],
                  "properties": {
                    "entrypoint": {
                      "type": "string"
                    },
                    "production": {
                      "type": "boolean"
                    },
                    "services": {
                      "type": "array",
                      "items": {
                        "anyOf": [
                          {
                            "type": "object",
                            "required": [
                              "id",
                              "type",
                              "status",
                              "version",
                              "localUrl",
                              "entrypoint",
                              "dependencies"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              },
                              "version": {
                                "type": "string"
                              },
                              "localUrl": {
                                "type": "string"
                              },
                              "entrypoint": {
                                "type": "boolean"
                              },
                              "workers": {
                                "type": "number"
                              },
                              "url": {
                                "type": "string"
                              },
                              "dependencies": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "required": [
                                    "id",
                                    "url",
                                    "local"
                                  ],
                                  "properties": {
                                    "id": {
                                      "type": "string"
                                    },
                                    "url": {
                                      "type": "string"
                                    },
                                    "local": {
                                      "type": "boolean"
                                    }
                                  }
                                }
                              }
                            }
                          },
                          {
                            "type": "object",
                            "required": [
                              "id",
                              "status"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/runtimes/{pid}/openapi/{serviceId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "pid",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "serviceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/runtimes/{pid}/restart": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "pid",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}